dev-requirements.txt
black
ruff
pytest>=8.0
pytest-cov
mypy
pre-commit
frappe-bench
isort==5.12.0
requests==2.31.0
click==8.1.7
qrcode
zxcvbn-python
